requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.0.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.8.0rc2
distlib==0.3.8
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-rest-swagger==2.2.0
django-simple-history==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dulwich==0.21.7
facebook-sdk==3.1.0
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.49.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
inflection==0.5.1
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keyring==24.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mistune==3.0.2
more-itertools==10.2.0
msgpack==1.0.7
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
openapi-codec==1.3.2
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==4.2.0
plotly==5.19.0
poetry==1.8.1
poetry-core==1.9.0
poetry-plugin-export==1.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rest-pandas==1.1.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.1.1
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
statsmodels==0.14.1
tenacity==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trove-classifiers==2024.2.23
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
webencodings==0.5.1
wheel==0.42.0
